Fixed buggy formatting on registration and login and splash and added quiz pictures name instead of placeholder.
