+++ education_app/pubspec.yaml
@@ -40,7 +40,6 @@ dependencies:
   cloud_firestore: ^4.15.6
   google_fonts: ^6.1.0
 
-
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -64,9 +63,10 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
+  assets:
+    - assets\images\quiz_logo.png
+    - assets\images\quiz_app_logo_3.png
+    - assets\images\quiz_app_logo_2.png
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
